refactor(terminal): #373 Phase 3 terminal.rs から race 無関係 helper を sub-module に move
Issue #373 / Phase 3 (PTY 境界整理): `commands/terminal.rs` (866 行) から
PTY race と無関係な純関数群を `commands/terminal/` 配下の 3 sub-module に分離。
move only / bit identical / IPC コマンド名・event 名は不変。

# 構成

```
commands/terminal.rs              866 → 486 行 (thin wrapper + race-sensitive のみ)
commands/terminal/
  paste_image.rs        新規  147 行  (extension_for_mime / cleanup_old_paste_images
                                       / MAX_PASTED_IMAGE_BYTES / save / mime_ext_tests)
  codex_instructions.rs 新規   51 行  (prepare_codex_instructions_file
                                       / cleanup_old_codex_instructions)
  command_validation.rs 新規  215 行  (is_valid_terminal_id / command_basename
                                       / configured_terminal_commands
                                       / is_allowed_terminal_command
                                       / reject_immediate_exec_args / is_codex_command
                                       + terminal_id_validation_tests / codex_command_tests)
```

# 不変式の維持

- `#[tauri::command] terminal_save_pasted_image` は thin wrapper として `terminal.rs`
  に残し、本体は `paste_image::save` に委譲 → `lib.rs` の `invoke_handler!` 登録
  および IPC コマンド名・event 名は **1 行も変えない**
- `SavePastedImageResult` struct は terminal.rs に残置 (Tauri IPC で公開する型)
- `inject_codex_prompt_to_pty` (race-sensitive、1.8s + 15ms チャンク) は touch せず
- `terminal_create` の id 衝突リトライ / `attach_if_exists` preflight / `claude_watcher`
  spawn 経路もすべて touch せず
- `use-pty-session.ts` も touch せず (引き継ぎ書 Phase 3 設計通り)

# 検証結果

- `cargo check`: ✅ green
- `cargo test --lib`: ✅ 83 passed / 0 failed (移動した 13 test も pass)
  - `commands::terminal::paste_image::mime_ext_tests::*` (1)
  - `commands::terminal::command_validation::terminal_id_validation_tests::*` (10)
  - `commands::terminal::command_validation::codex_command_tests::*` (2)
- `cargo clippy --no-deps`: ✅ 13 warnings = baseline 15 件 - 2 件
  (#10 protocol.rs:881 は Phase 2 分割で消滅 / #12 mod.rs:488 too_many_arguments も消滅)
  新規 lint 種類・新規 root cause は 0 件。`command_validation.rs:57` の
  doc_lazy_continuation は baseline #5 (元 `terminal.rs:136`) が move 先に追従しただけで
  bit identical の必然 (将来 `tasks/refactor-clippy-baseline.md` の解消 PR で巻き取り対象)

# Agent Team で並列分担

3 sub-module の draft を 3 人の teammate (paste-image / codex-instructions
/ command-validation extractor) に並列で書かせ、team-lead 側で統合 + cargo check
+ thin wrapper 化 + clippy 通しを実施。

Refs #373
